requirements.txt
# Core dependencies for Custom AI Agent Meeting Scheduler
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# AI/LLM Integration (Gemini 2.0 Flash)
google-generativeai>=0.3.0

# Enhanced CLI components
rich>=13.0.0
tabulate>=0.9.0

# Development and testing
pytest>=7.0.0

# Environment configuration
python-dotenv>=1.0.0
